URL restriction for unwanted user part done

+++ base/views.py
@@ -1,7 +1,9 @@
 from django.contrib.auth import authenticate
 from django.contrib.auth.models import User
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
@@ -9,6 +11,9 @@ from .forms import RoomForm
 
 
 def loginPage(request):
+
+  if request.user.is_authenticated:
+    return redirect('home')
 
   if request.method == 'POST':
     username = request.POST.get('username')
@@ -54,6 +59,8 @@ def room(request, pk):
   context = {'room':room}
   return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')#decorater if the session id in not in the browser they will now show the createroom button
 def createRoom(request):
   form = RoomForm()
 
@@ -66,10 +73,13 @@ def createRoom(request):
   context = {'form':form}
   return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
   room = Room.objects.get(id=pk)
   form = RoomForm(instance=room)
+
+  if request.user != room.host:
+    return HttpResponse('you are not allowed here!')
 
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
@@ -81,8 +91,13 @@ def updateRoom(request, pk):
   return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
   room = Room.objects.get(id=pk)
+
+  if request.user != room.host:
+    return HttpResponse('you are not allowed here!')
+
   if request.method == 'POST':
     room.delete()
     return redirect('home')
